上传接口修改和数据库脚本备份

+++ sexy-tea/src/main/java/sexy/tea/controller/CardController.java
@@ -28,7 +28,7 @@ public class CardController {
     }
 
     @GetMapping("/selectByUid/{uid}")
-    public Result selectByUid(@PathVariable Integer uid) {
+    public Result selectByUid(@PathVariable Long uid) {
         return service.findByUid(uid);
     }
