+++ sexy-tea/src/main/java/sexy/tea/service/BeverageService.java
@@ -24,11 +24,11 @@ public interface BeverageService extends UploadService {
 
     Result find(int pageNum, int pageSize);
 
-    Result findByPrimaryKey(Integer primaryKey);
+    Result findByPrimaryKey(Long primaryKey);
 
     Result saveOrUpdate(Beverage beverage);
 
-    Result delete(Integer id);
+    Result delete(Long id);
 
     Result findByName(String name, int pageNum, int pageSize);
 }
